Create node system for navigation
Create node system for player and enemies to move along, and for coins to be places on.
Node.isTraversable indicates that the player/enemies can move along the node.
Node.canPlacePellet indicates that a coin/pellet can be place on it.
Each node uses OverlapCircleAll to check for the layer mask of an overlapping obstacle (Wall layer -> !isTraversable, Empty layer -> !canPlacePellet).